Adds cve to cwe mapper
